pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sample-editor"
version = "0.1.0"
description = "Desktop-friendly revamp of SampleEditor"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "imagecodecs",
  "numpy",
  "pillow",
  "qrcode[pil]",
  "tifffile",
]

[project.optional-dependencies]
gui = [
  "PySide6",
]
dev = [
  "pytest",
]

[project.scripts]
sample-editor = "sample_editor.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
